Plan: plans/PR-H18-Watchlist-Alert-Events-Source-Attestation.md
fix: isolate watchlist evaluator from application imports

Move shared alert helpers into the service and defer delivery-only imports until send/render operations. The migration proof now imports the real evaluator with only its workflow dependency set, while dashboard helper behavior remains unchanged.
